[refactor] test code
