+++ src/test/java/flab/project/common/FileStorage/BucketUtilsTest.java
@@ -1,6 +1,6 @@
 package flab.project.common.FileStorage;
 
-import flab.project.data.enums.FileType;
+import flab.project.domain.file.enums.FileType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
