+++ src/test/java/flab/project/controller/LikeControllerTest.java
@@ -3,7 +3,8 @@ package flab.project.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import flab.project.config.baseresponse.ResponseEnum;
 import flab.project.config.baseresponse.SuccessResponse;
-import flab.project.service.LikeService;
+import flab.project.domain.like.controller.LikeController;
+import flab.project.domain.like.service.LikeService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
