+++ src/test/java/flab/project/controller/PostControllerTest.java
@@ -2,13 +2,14 @@ package flab.project.controller;
 
 import flab.project.config.baseresponse.ResponseEnum;
 import flab.project.config.baseresponse.SuccessResponse;
-import flab.project.data.dto.PostWithUser;
-import flab.project.data.dto.model.BasicPost;
-import flab.project.data.dto.model.BasicUser;
-import flab.project.data.dto.model.DebatePost;
-import flab.project.data.dto.model.PollPost;
-import flab.project.data.enums.PostType;
-import flab.project.service.PostService;
+import flab.project.domain.post.model.PostWithUser;
+import flab.project.domain.post.model.BasicPost;
+import flab.project.domain.user.model.BasicUser;
+import flab.project.domain.post.model.DebatePost;
+import flab.project.domain.post.model.PollPost;
+import flab.project.domain.post.enums.PostType;
+import flab.project.domain.post.controller.PostController;
+import flab.project.domain.post.service.PostService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +22,7 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.ResultMatcher;
 
 import static flab.project.config.baseresponse.ResponseEnum.*;
-import static flab.project.data.enums.PostType.*;
+import static flab.project.domain.post.enums.PostType.*;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
