+++ src/test/java/flab/project/controller/VoteControllerTest.java
@@ -4,8 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import flab.project.config.baseresponse.ResponseEnum;
 import flab.project.config.baseresponse.SuccessResponse;
 import flab.project.config.exception.NotFoundException;
-import flab.project.data.enums.PostType;
-import flab.project.service.VoteService;
+import flab.project.domain.post.enums.PostType;
+import flab.project.domain.post.controller.VoteController;
+import flab.project.domain.post.service.VoteService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
