+++ src/test/java/flab/project/data/dto/UpdateSocialAccountRequestDtoTest.java
@@ -1,10 +1,10 @@
 package flab.project.data.dto;
 
+import flab.project.domain.user.model.UpdateSocialAccountRequestDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class UpdateSocialAccountRequestDtoTest {
