+++ src/test/java/flab/project/data/dto/model/SocialAccountTest.java
@@ -1,11 +1,11 @@
 package flab.project.data.dto.model;
 
-import flab.project.data.dto.UpdateSocialAccountRequestDto;
+import flab.project.domain.user.model.SocialAccount;
+import flab.project.domain.user.model.UpdateSocialAccountRequestDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class SocialAccountTest {
