+++ src/test/java/flab/project/facade/UserFacadeTest.java
@@ -4,8 +4,9 @@ import flab.project.common.FileStorage.BucketUtils;
 import flab.project.common.FileStorage.FileExtensionFilter;
 import flab.project.common.FileStorage.FileStorage;
 import flab.project.config.exception.FailedToUpdateProfileImageToDatabaseException;
-import flab.project.data.enums.FileType;
-import flab.project.service.UserService;
+import flab.project.domain.file.enums.FileType;
+import flab.project.domain.user.facade.UserFacade;
+import flab.project.domain.user.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -18,7 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.ArrayList;
 import java.util.List;
 
-import static flab.project.data.enums.FileType.PROFILE_IMAGE;
+import static flab.project.domain.file.enums.FileType.PROFILE_IMAGE;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.given;
