+++ src/test/java/flab/project/service/IconServiceTest.java
@@ -1,18 +1,17 @@
 package flab.project.service;
 
-import flab.project.data.dto.UpdateSocialAccountRequestDto;
-import flab.project.data.dto.model.SocialAccount;
-import flab.project.mapper.IconMapper;
+import flab.project.domain.user.model.UpdateSocialAccountRequestDto;
+import flab.project.domain.user.model.SocialAccount;
+import flab.project.domain.user.service.IconService;
+import flab.project.domain.user.mapper.IconMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
